apply eclipse format and clean up

+++ src/main/java/de/tud/stg/analysis/ObjectTrace.java
@@ -2,15 +2,14 @@ package de.tud.stg.analysis;
 
 import java.util.*;
 
-import org.apache.commons.lang.StringUtils;
 /**
- * 		//Note that since the relationship holds for the identity, i.e. xEx is always-
-		//valid, E(x) always contains x itself, and |E(x)|  1.
-
-		// well, this reuce the average S-score significantly
-		// we remove the 1 again
-
-
+ * //Note that since the relationship holds for the identity, i.e. xEx is
+ * always- //valid, E(x) always contains x itself, and |E(x)| 1.
+ * 
+ * // well, this reuce the average S-score significantly // we remove the 1
+ * again
+ * 
+ * 
  * @author martin
  *
  */
@@ -20,25 +19,29 @@ public class ObjectTrace extends TypeUsage {
 	public int nequals;
 	public int nalmostequals;
 	public static final double conservative_coef = 2;
+
 	public void reset() {
-		nequals=0;
-		nalmostequals=0;
+		nequals = 0;
+		nalmostequals = 0;
 	}
-	
+
 	public double strangeness() {
-		if ((nequals+nalmostequals)==0) {
+		if ((nequals + nalmostequals) == 0) {
 			return 0;
 		}
-		return ((double)nalmostequals)/(nequals+nalmostequals); 
-	}		
+		return ((double) nalmostequals) / (nequals + nalmostequals);
+	}
+
 	public double strangeness2() {
 		return strangeness2(conservative_coef);
 	}
+
 	public double strangeness2(double coef) {
-		return (nalmostequals)/((double)(Math.pow(nequals, coef)+nalmostequals));			
+		return (nalmostequals) / (Math.pow(nequals, coef) + nalmostequals);
 	}
+
 	public double strangeness3() {
-		return ((double)nalmostequals)/(1+ nequals +nalmostequals);			
+		return ((double) nalmostequals) / (1 + nequals + nalmostequals);
 	}
 
 }
